+++ Bolnica/vezba/SecretaryGUI/SecretaryViewDoctor.xaml.cs
@@ -30,7 +30,7 @@ namespace vezba.SecretaryGUI
             InitializeComponent();
             Doctor = selectedDoctor;
             DoctorService doctorService = new DoctorService();
-            
+
 
             WorkingHours = new ObservableCollection<WorkingHours>(doctorService.GetFutureWorkingHoursForDoctor(Doctor.Jmbg));
             VacationDays = new ObservableCollection<VacationDays>(doctorService.GetFutureVacationDaysForDoctor(Doctor.Jmbg));
@@ -67,6 +67,12 @@ namespace vezba.SecretaryGUI
             if (workingScheduleTable.SelectedCells.Count > 0)
             {
                 WorkingHours selectedWorkingHours = (WorkingHours)workingScheduleTable.SelectedItem;
+                SecretaryDeleteConfirmation dc = new SecretaryDeleteConfirmation(Doctor, selectedWorkingHours);
+                Boolean ic = false;
+                dc.ShowDialog();
+                ic = dc.isConfirmed;
+                if (!ic)
+                    return;
                 DoctorService doctorService = new DoctorService();
                 doctorService.RemoveWorkingHoursFromDoctor(Doctor.Jmbg, selectedWorkingHours);
                 WorkingHours.Remove(selectedWorkingHours);
@@ -77,7 +83,7 @@ namespace vezba.SecretaryGUI
 
         private void AddVacationDaysButton_Click(object sender, RoutedEventArgs e)
         {
-           SecretaryNewVacationDays w = new SecretaryNewVacationDays(Doctor);
+            SecretaryNewVacationDays w = new SecretaryNewVacationDays(Doctor);
             w.Show();
         }
 
@@ -86,12 +92,42 @@ namespace vezba.SecretaryGUI
             if (vacationTable.SelectedCells.Count > 0)
             {
                 VacationDays selectedVacationDays = (VacationDays)vacationTable.SelectedItem;
+                SecretaryDeleteConfirmation dc = new SecretaryDeleteConfirmation(Doctor, selectedVacationDays);
+                Boolean ic = false;
+                dc.ShowDialog();
+                ic = dc.isConfirmed;
+                if (!ic)
+                    return;
                 DoctorService doctorService = new DoctorService();
                 doctorService.RemoveVacationDaysFromDoctor(Doctor.Jmbg, selectedVacationDays);
                 VacationDays.Remove(selectedVacationDays);
             }
             else
                 MessageBox.Show("Niste selektovali godisnji odmor!");
+        }
+
+        private void WindowKeyListener(object sender, KeyEventArgs e)
+        {
+            if (e.Key == Key.Escape)
+                this.Close();
+            else if (e.Key == Key.Enter)
+                this.Close();
+        }
+
+        private void OnKeyDownDataGrid1Handler(object sender, KeyEventArgs e)
+        {
+            if (e.Key == Key.A)
+                this.AddWorkingHoursButton_Click(sender, e);
+            else if (e.Key == Key.D)
+                this.RemoveWorkingHoursButton_Click(sender, e);
+        }
+
+        private void OnKeyDownDataGrid2Handler(object sender, KeyEventArgs e)
+        {
+            if (e.Key == Key.B)
+                this.AddVacationDaysButton_Click(sender, e);
+            else if (e.Key == Key.G)
+                this.RemoveVacationDaysButton_Click(sender, e);
         }
     }
 }
